+++ aquilia/cli/compilers/__init__.py
@@ -1,7 +1,0 @@
-"""Artifact compilers."""
-
-from .workspace import WorkspaceCompiler
-
-__all__ = [
-    "WorkspaceCompiler",
-]